Goggles vs Screen synchronization diagnostics
Check if screen and goggles are synchronized (subjectively) with a variety of goggle delays. Goggles also now open after (and not before) the Screen('Flip') command in the script, which may help synchronization. Due to monitor display lag, the stimulus tends to onset slightly after the goggles, which have a response time of around 5ms.